backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.6
djangorestframework==3.14.0
exceptiongroup==1.1.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
kombu==5.2.4
mccabe==0.7.0
packaging==23.0
pep8-naming==0.13.3
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
